Use the outbound fifo (THR)
The THR_EMPTY flag is set when the THR is empty.  When this happens, the
ISR has a loop to load data from the TX buffer into the THR, but the
loop terminates when the THR_EMPTY flag is clear.  Since this happens
after the first byte, it can never push more than one character per IRQ.

Instead, the ISR can trust the THR_EMPTY condition and try to move up to
the limit (16 bytes) from the TX buffer into the THR.